+++ templates/vue/src/components/Lightbox/media/ActivityMedia/Question/TextQuestion.vue
@@ -1,27 +1,56 @@
 <template>
   <b-form class="grid-container" @submit="handleTextSubmit">
-    <p class="form-input">
-      <b-form-textarea
-        v-if="question.answerTypes.text.isMultiLine"
-        v-model="text"
-        rows="5"
-      ></b-form-textarea>
-      <b-form-input
-        v-else
-        v-model="text"
-        :placeholder="question.answerTypes.text.placeholder"
-      ></b-form-input>
-      <b-form-invalid-feedback :state="isAnswerValid">
-        Please enter a response.
-      </b-form-invalid-feedback>
-      <b-button class="mt-3 float-right" variant="primary" type="submit">
-        Submit
-      </b-button>
-    </p>
+    <ul v-if="question.answerTypes.text.allowMultiple" class="list">
+      <li v-for="(answerItem, index) in textAnswers" :key="index" class="answerItem">
+        <b-form-input
+          v-model="textAnswers[index]"
+          :data-qa="`list-input-${index}`"
+          :placeholder="
+            question.answerTypes.text.placeholder
+              ? question.answerTypes.text.placeholder
+              : 'Type an answer and press Enter'
+          "
+        ></b-form-input>
+        <b-button
+          v-show="numOfAnswers < maxFields"
+          variant="primary"
+          :data-qa="`list-add-${index}`"
+          @click="addAnswer"
+          @keydown.enter="enterPress"
+        >
+          +
+        </b-button>
+        <b-button
+          v-show="numOfAnswers > minFields"
+          variant="danger"
+          :data-qa="`list-del-${index}`"
+          @click="deleteAnswer(index)"
+        >
+          -
+        </b-button>
+      </li>
+    </ul>
+    <b-form-textarea
+      v-else-if="question.answerTypes.text.isMultiLine"
+      v-model="textAnswers[0]"
+      rows="5"
+    ></b-form-textarea>
+    <b-form-input
+      v-else
+      v-model="textAnswers[0]"
+      :placeholder="question.answerTypes.text.placeholder"
+    ></b-form-input>
+    <b-form-invalid-feedback :state="isAnswerValid">
+      Please enter a response.
+    </b-form-invalid-feedback>
+    <b-button class="mt-3 float-right" variant="primary" type="submit">
+      Submit
+    </b-button>
   </b-form>
 </template>
 
 <script>
+const MAX_FIELDS_ALLOWED = 100
 export default {
   name: "text-question",
   props: {
@@ -30,38 +59,65 @@ export default {
       required: true,
     },
     answer: {
-      type: [Array, String],
+      type: Array,
       required: true,
     },
   },
   data() {
     return {
       isAnswerValid: true,
-      text: "",
+      textAnswers: this.answer.map(x => x),
     }
   },
-  watch: {
-    text(newText) {
-      this.text = newText
+  computed: {
+    minFields() {
+      return parseInt(this.question.answerTypes.text.list.minFields, 10)
     },
-    question() {
-      this.text = this.answer
+    maxFields() {
+      return this.question.answerTypes.text.list.maxFields.enabled
+        ? parseInt(this.question.answerTypes.text.list.maxFields.value, 10)
+        : MAX_FIELDS_ALLOWED
+    },
+    numOfAnswers() {
+      return this.textAnswers.length
     },
   },
   created() {
-    if (Array.isArray(this.answer)) {
-      this.text = this.answer[0]
-    } else {
-      this.text = this.answer
+    if (this.question.answerTypes.text.allowMultiple) {
+      for (let i = this.numOfAnswers; i < this.minFields; i++) {
+        this.addAnswer()
+      }
+    }
+    if (this.answer.length > 1 && !this.question.answerTypes.text.allowMultiple) {
+      this.textAnswers = []
+      this.textAnswers.push(this.answer.join())
     }
   },
   methods: {
     handleTextSubmit(event) {
       event.preventDefault()
-      this.isAnswerValid = this.text !== ""
+      this.isAnswerValid = this.textAnswers[0] !== ""
       if (this.isAnswerValid) {
-        this.$emit("submit", this.text)
+        this.$emit("submit", this.textAnswers)
       }
+    },
+    deleteAnswer(index) {
+      this.$bvModal
+        .msgBoxConfirm("This answer will be removed.", {
+          modalClass: "node-modal-confirmation",
+          title: "Are you sure you want to delete this answer from the list?",
+          okTitle: "Yes, delete!",
+          okVariant: "danger",
+        })
+        .then(close => {
+          if (close) {
+            this.textAnswers.splice(index, 1)
+          }
+        })
+        .catch(err => console.log(err))
+    },
+    addAnswer() {
+      this.textAnswers.push("")
     },
   },
 }
@@ -71,6 +127,34 @@ export default {
 .grid-container {
   display: grid;
   align-content: center;
-  height: 250px;
+}
+
+.answerItem {
+  display: flex;
+  border-style: solid;
+  border-radius: 6px;
+  padding: 10px 20px;
+  margin-top: 5px;
+}
+
+.list li button {
+  float: right;
+  position: relative;
+  padding: 20px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  height: 20px;
+  width: 50px;
+  float: right;
+  font-size: 30px;
+  font-weight: bold;
+  margin-left: 3px;
+  margin-right: 3px;
+}
+
+.submit-btn {
+  float: center;
+  margin-top: 30px;
 }
 </style>
